[Parser] mode command and type parsing
